Separate code example renderers into separate files.
