+++ lib/hologram/code_example_renderers.rb
@@ -1,0 +1,8 @@
+module Hologram
+  module CodeExampleRenderers
+  end
+end
+
+Dir[File.join(File.dirname(__FILE__), 'code_example_renderers', '*')].each do |file|
+  require file
+end
