+++ lib/hologram/code_example_renderers/js_example.rb
@@ -1,0 +1,12 @@
+module Hologram
+  module CodeExampleRenderers
+    class JsExample < Example
+      private
+
+      def lexer
+        @_lexer ||= Rouge::Lexer.find('js')
+      end
+    end
+  end
+end
+
