+++ lib/hologram/code_example_renderers/jsx_example.rb
@@ -1,0 +1,12 @@
+module Hologram
+  module CodeExampleRenderers
+    class JsxExample < Example
+      private
+
+      def lexer
+        @_lexer ||= Rouge::Lexer.find('html')
+      end
+    end
+  end
+end
+
